pyproject.toml
[project]
name = "fraudguard"
version = "0.1.0"
description = "Production-ready GNN fraud detection with RL subgraph selection"
authors = [{name = "Jose A"}]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
    # NOTE: torch-scatter and torch-sparse need special installation:
    # pip install torch-scatter torch-sparse -f https://data.pyg.org/whl/torch-{VERSION}+{CUDA}.html
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "faiss-cpu>=1.7.4",
    "pyyaml>=6.0",
    "structlog>=23.1.0",
    "mlflow>=2.10.0",
    # API dependencies
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
